refactor: controller, domain 패키지 분리
