+++ src/main/java/roomescape/reservation/controller/AdminViewController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.reservation.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
