+++ src/main/java/roomescape/reservation/controller/response/ReservationResponse.java
@@ -1,8 +1,9 @@
-package roomescape;
+package roomescape.reservation.controller.response;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
+import roomescape.reservation.domain.Reservation;
 
 public record ReservationResponse(
         Long id,
